some changes

+++ client/src/components/WorkoutDetails.jsx
@@ -1,15 +1,24 @@
 import React from "react";
 import { useWorkoutContext } from "../hooks/useWorkoutContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutContext();
 
+  const {user} = useAuthContext();
+
   const handleClick = async () => {
+    if (!user) {
+      return
+    }
     const response = await fetch(
       `http://localhost:4000/api/workouts/${workout._id}`,
       {
         method: "DELETE",
+        headers : {
+          "Authorization": `Bearer ${user.token}`
+        }
       }
     );
     const data = await response.json();
